Added: Hexgonal Architecture Django docs
