+++ had/app/management/commands/create_password.py
@@ -7,7 +7,7 @@ from django.contrib.auth.hashers import make_password
 
 class Command(BaseCommand):
 
-    help = 'Create users password'
+    help = 'Create user password'
 
     def add_arguments(self, parser):
         parser.add_argument('-p', '--password',
