+++ src/shared/infrastructure/persistence/django/__init__.py
@@ -1,7 +1,7 @@
 from .django_repository import DjangoRepository
-from .django_orm_manager import DjangoOrmManager
+
 
 __all__ = (
-    'DjangoRepository', 'DjangoOrmManager',
+    'DjangoRepository',
 )
