+++ src/shared/infrastructure/persistence/django/django_repository.pyi
@@ -1,12 +1,1 @@
-from src.shared.domain.repository import AbstractRepository
-
-
-class DjangoRepository(AbstractRepository):
-
-    def add(self, entity): ...
-
-    def edit(self, entity): ...
-
-    def delete(self, entity): ...
-
-    def search(self, **fields) -> : ...
+class DjangoRepository: ...
