+++ src/users/domain/repository/__init__.py
@@ -1,0 +1,4 @@
+from .user_repository import UserRepository
+
+
+__all__ = ('UserRepository', )
