+++ src/users/domain/repository/user_repository.py
@@ -1,0 +1,8 @@
+from abc import ABC, abstractmethod
+
+
+class UserRepository(ABC):
+    @abstractmethod
+    def all(self):
+        raise NotImplementedError("Not implemented error")
+
